Rename identity_commiter -> task_monitor
